ci: prune unpushed images before acquisition
This fixes one of my longest standing gripes with CI, which is that when
all images already exist (e.g. because no code has changed), the build
job still has to download all the existing images from Docker Hub.

This commit reworks the build job so that the only images downloaded are
those that are dependencies of other images that need to be built.