materialized: support Materialize Cloud authentication
Add a new cli flag to specify a tenant and JWK causing pgwire and the
http endpoint to require a valid JWT from Materialize Cloud. We trust
the validated username, and will create it if it does not exist.

The JWT comes with expiry, refresh, and role information that we currently
ignore, but will use in the future. The cloud sdk has a function that
does some of this, but omits the expiry information, so isn't good enough
for use here.