ci/test: make the job names consistent
The emojis were confusing, because they were included in the GitHub
check names like "paperclip-clippy-and-doctests". This only makes sense
for the tests with a truly relevant emoji, like ":csharp: tests".

Also take the opportunity to standardize on phrasing, so we always say
"Foo smoke test" or "Foo tests" rather than e.g. "Tests for the foo".