remove concept of original ID from sources
This is a leftover from when sources were made of two parts, a bare
source and a mini view that post processed the results of the bare
source.

Since the decoupling of debezium from decoding this is no longer the
case and sources always directly output their final collection.

`orig_id` was previously used to differenticate between the bare source
and the mini view but now the ids are always the same, so we can remove
the concepct of orignal ID altogether.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>