sources: remove BYO timestamping
BYO timestamping allowed users to provide their own mapping between
message offsets in a kafka data topic and the timestamp they get
assigned. This mechanism could be used to preserve transaction semantics
by ensuring that all relevant events of a transaction happen at the same
timestamp.

To the best of my knowledge no one is actually using this feature and it
has never been mentioned in our documentation. The mechanism is also
specific to Kafka and doesn't lend itself very well to other source
types. Its main use until now has been in our testdrive test suite where
it served as a way to very precicely control the timestamping of
ingested data.

The motivation for removal of this feature is its entangled nature with
the coordinator which currently has code to connect to kafka, decode
avro, and a hardcoded schema of the debezium transaction schema, all
part of what is known as the timestamper. As part of platform we want to
move all the timestamper logic out of the coordinator and keeping BYO
support would make this goal ten times more difficult.

For the above reasons we'll be dropping support for BYO timestamping.
All the tests that were affected by this have been rewritten to use our
`MATERIALIZE` envelope for controlling timestamps which is much easier
to work with and supports sending both data and progress events inline
in the same stream.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>