coord: refactor handling of source persist details
The primary change in this PR is that `SourcePersistDesc` struct is
transmitted to the dataflow layer by way of the `SourceInstanceDesc`
struct, rather than via injection into the `SourceConnector` struct.

This clarity fixes several unreported bugs:

  * The `persist_details` field of mz_sources is correctly populated.
  * The `since_ts` of recreated sources is correctly restored.

Both of these bugs were the result of relying on the persist details in
the `SourceConnector` before those details were injected, so the details
were always `None`.