persist: fix flaky conditional_seal test
Previously, the conditional test would very rarely fail because the primary stream
had sealed to a time (4) that we did not explicitly ask it to.

This happened because:
- the timely::execute_directly command on drop closes all unclosed inputs and
  runs the dataflow to completion.
- Progress information (including closure) doesn't travel through the dataflow
  all at the same time, even if two inputs are both closed before calling step()

This lead to a situation where if the condition input had been advanced (and sealed)
through 4, and the primary input had been advanced and sealed through 3, the
primary seal operator (which sits downstream of both sources) could observe the
primary input's closure (ie that the primary input has the frontier []), before
it observes the condition input's closure and thereby conclude that its frontier
is 4, and seal primary to 4.

The fix in this commit is simple: we directly drop all dataflows installed for
the test before returning from timely::execute_directly.

This does not seem like an issue outside of this test for the following reasons:
1. Materialize doesn't (yet) have a graceful shutdown path. Our shutdown consists
   of running Ctrl-C, so there's no actual closing of dataflows and waiting for
   the dataflow to be fully closed.
2. Even if that were not the case, the "problem" here is that primary sealed 4, when
   the test did not expect it. One could argue that the test's expectations here
   are perhaps too rigid, and that after closing an input persist is allowed
   to seal the corresponding persistent collection to any unsealed times.

Other potential fixes: this is the rare case where we could have avoided a problem
by writing a custom dataflow operator for conditional seal, rather than using the
flat_map(|_| None) trick to induce an artificial data dependency.