dbt-materialize: fix list_relations_without_caching
Previously, as far as I could tell, this macro always returned an empty
list. Fix the bug by importing the upstream Postgres's definition of the
macro, rather than reimplementing it ourselves.

This depends on `pg_catalog` and `information_schema` views that will be
introduced in the next release (v0.20.0), so update the minimum
supported Materialize version accordingly.