debezium: handle out of order upsert updates
Previously out upsert implementation assumed that all updates for a
particular key will arrive in order. However there is no guarantee that
this will happen and indeed recent changes to the decoding pipeline
caused our upsert logic to fail.

This patch adds explicit tracking of the last offset seen for each
key that in the upsert state and rejects any upsert message that is in
the past.

Fixes #6715

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>