sql: handle type conversions in set expressions
Apply type conversion rules to the LHS and RHS of a UNION, INTERSECT, or
EXCEPT operation. This matches PostgreSQL and improves compatibility
with dbt, which does a lot of set operations on tables with compatible
but not identical types.

This fixes most of #3333. Still outstanding is the handling of coercible
expressions (i.e., untyped string literals) as described in that issue.

The tests in this patch are adapted from PostgreSQL. They test set
operations generally, not just the mixed type operations that this patch
adds support for.