retire timestamping thread from coordinator
 ## Overview

This patch aims to simplify how timestamp assignment is handled in the
system by removing the timestamping thread hosted by the coordinator and
deleting the auxiliary `ConsistencyInfo` and `ConsInfo` data structures
from the source handling code.

 ## High level description of changes

 ### Partition discovery

The first thing that this commit changes is how partition discovery
works. Previously the timestamper thread in the coordinator was being
notified every time a source was being created (regardless of whether or
not that source had been instanciated yet in a dataflow). The
timestamper thread would react to that by starting monitoring for
partitions. In theory the notion of a partition could exist for any type
of source but in practive only Kafka sources utilize the notion of a
partition.

Whenever the timestamper thread discovered a new partition it sent a
message to the coordinator with the new ids and the coordinator was
subsequently broadcasting this information to all the dataflow workers
through the `AdvanceSourceTimestamp` command.

The timestamping thread, the `AdvanceSourceTimestamp` command, and all
associated machinery have been removed. Source implementations are now
responsible for discovering partitions themselves.

**Note**: The implementation of this PR is a bit wasteful in that every
single worker for every single source instance will be polling for
partitions. While the code is carefully written to make sure each
partition is only handled by a single worker it would be nice to have
this be done centrally and broadcasted to each worker. This falls under
the future work of building a Source trait that encompaces this pattern.

 ### Timestamp binding management

Both before and after this commit the authoritative source for which
timestamp bindings have been minted is the `TimestampBindingRc`
structure. This structure is constracted once when a source gets created
(via the `AddSourceTimestamping` dataflow command) and is shared between
all the sources instances that will be created in the future.

Ultimately the goal of timestamp assignment is to get two things right.
One is to make sure the `handle_message` function selects the correct
timestamp for the message it is currently processing, and for the
operator to downgrade its capabilities to the right timestamps.

The main difference between the state of affairs before and after this
commit is that previously both of the above tasks were performed through
a series of intermetiate objects (`ConsistencyInfo` and `ConsInfo`) that
served as a cache of information. In the past they were also responsible
for a lot of the BYO timestamping code.

Since BYO timestamping is no longer a thing we can massively simplify
the interaction of the source timely operator and the TimestampBindingRc
structure. Specifically, the source only ever calls the
`get_or_propose_timestamp(partition, offset)` method which will always
give back a timestamp. This timestamp will either come from some already
minted binding or will be assigned the currently open timestamp of the
proposer.

Before this commit it was possible that `TimestampBindingRc` did not
produce a timestamp (i.e was returning an Option). This was only
happening in the BYO timestamping case so again the code has been
simplified to always return a timestamp.

 ### Changes to SourceReader trait

The `SourceReader` trait used to have an `add_partition` method that
panicked by default and was only implemented by Kafka. Since all sources
are responsible for discovering partitions now this method has been
removed.

Furthermore, the `SourceReader::new` method has gained two new
arguments. One is `worker_count`, a piece of information complementing
the `worker_id` that was already there and is used to determine which
worker is responsible for which partition.

The other one is `restored_offsets` which is used when a source is
instanciated after we have already persisted some of it. This currently
happens only when experimental features are enabled and we are
recovering from a crash (or we're simply restarting). Source
implementations must use this information to skip ahead to the specified
offsets before emitting data. This information was previously
communicated to sources via the `add_partition` method.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>