Add support for AWS External IDs when assuming roles
This adds the ability for Materialize Cloud to specify external IDs via the
command line. Users of Materialize Cloud will never be able to specify an
external ID, which is an important property of External IDs -- they are not
secrets, they are tags that providers specify and users do not.