test/kafka-ingest: remove demand optimization when rendering upsert sources
Previously, upsert sources used much less memory than expected on the open loop
benchmark. For example, after sending ~20 GiB, we would expect materialize's
memory usage with upsert sources to be approximately 20 GiB, if not more. Instead,
the memory usage was more like 8 GiB. This happened because the views built on
top of the upsert source didn't actually use the data in any of the columns
so the upsert logic replaced all of the columns with placeholder empty datums.

This only becomes a problem when comparing persistent upsert sources with non
persistent upsert sources. Persistent upsert sources are not able to push
the fact that the columns are projected away all the way down into the upsert
logic, and consequently use a lot more memory than non-persistent upsert sources.
This ends up confounding the goal of measuring the memory overhead of the persist
runtime, as now we're not just comparing the effects of enabling vs disabling
persist, we're doing so across two very different dataflows.

A reasonable counterargument here might be something like: not being able to
push projections down as far into stateful logic is a direct consequence of using
persist, and thus we should be accounting for the overheads that result from
the lack of optimization. The best response to that line of thought is that
at the moment, we're more focused on the direct overheads of persist itself,
rather than the indirect costs of using persist. Furthermore, the demand optimization
only applies in the artificial case of the dataflow that throws away all of the
columns. We don't expect to see workloads like that from alpha testers, and if
we do, we can revisit pushing projections/MFPs into persistent upsert.