sql,sql-parser: hoist operator semantic analysis out of parser
We only support qualified operators when the namespace is `pg_catalog`.
This is semantic analysis, though, so belongs in the planner rather than
the parser. Restructure the operator parsing/planning code accordingly.

I also took the opportunity to simplify the `Op` AST node and the
`parse_operator` method. There's no need to use an enum because an empty
vector works just fine; and in fact the new representation avoids having
two separate ways to represent unqualified operators.