Support for LIKE escape strings, not just escape characters. (#10572)
In my PR #10055, I added the beginnings of support for user-specified escape characters. I realized later that the input to `ESCAPE` is a string, not a character, and that an empty string is valid and means that the escaping mechanism is disabled. This PR adds support for full escape strings to `like_pattern::compile()`.

The way it does this is through introducing a pattern normalization step, where any user-specified escape settings are used to convert the pattern to one that uses the default escape string. (`'\\'`) This is slightly less efficient, but works well with Materialize's diagnostic output. Explain plans and other debug output sometimes use the `~~` and `~~*` operators; these binary operators do not allow for user-supplied escape strings. Normalizing to the default escape string lets us display the stored pattern in the form that would be expected by these operators.

This pattern normalization does come with a cost, but it is minimal in the common case:

`like_pattern/like_compile`:
   before:   [227.77 ns **228.12 ns** 228.55 ns]
   after:    [233.13 ns **233.31 ns** 233.54 ns] (default escape string)
   after:   [280.09 ns **280.57 ns** 281.46 ns] (custom escape string)

`like_pattern/ilike_compile`:
   before:   [22.888 us **22.915 us** 22.951 us]
   after:    [23.346 us **23.573 us** 23.845 us]  (default escape string)
   after:    [23.505 us **23.608 us** 23.734 us]  (custom escape string)

For more information on escape strings, see the PostgreSQL documentation:
https://www.postgresql.org/docs/8.4/functions-matching.html