sql: use `ScalarType` rather than `mz_pgrepr::Type` as param type
The `mz_pgrepr::Type` is a PostgreSQL implementation detail but had
leaked into the Materialize-specific `sql` and `coord` crates. Correct
that boundary violation. The code to take a `mz_pgrepr::Value` into a
`Datum` gets much simpler in the process; the `pgrepr::Type` type is now
directly convertible to a `ScalarType` without looking at the datums
involved, so we can massively simplify the signature of
`Value::into_datum` and entirely remove `pgrepr::null_datum`.