coord,sql: store types differently in catalog
Store types in the catalog via a new CatalogType enum. This enum is
nearly one-to-one with ScalarType, except that modifier fields are
removed and embedded types are replaced with GlobalId references.

This type is used throughout the SQL planner, in the plans returned by
`CREATE TYPE` and in the type name resolution code.

The motivation for this refactor is the removal of the "lossy"
conversions from `pgrepr::Type` to `ScalarType`. Now, all conversions
are either full fidelity or report an error if they would discard data.