ci: increase the priority of longer-running jobs
Buildkite schedules jobs based on the order of their appearance
in the .yml file. This causes some longer-running jobs to wait
for an available agent for a while, which increases the total
execution time of the entire pipeline.

So, move certain longer-running jobs to the front, which will
cause them to be scheduled before all others.