expr,pgrepr: reuse logic from pgrepr in mz_render_typmod
The mz_render_typmod function recently became duplicative with the new
typmod rendering code in pgrepr. This commit refactors mz_render_typmod
to use reuse the logic from pgrepr, so that the logic lives in precisely
one place.

This commit also renames mz_render_typemod to mz_render_typmod to
consistently use the spelling "typmod" throughout the codebase. This is
an internal-only function, so the change does not require a release
note.

The changes to the types.slt reflect changes to numeric typemods in the
forthcoming PostgreSQL 15. See postgres/postgres@085f931f5.