truncate datums (#10767)
See comment in the code.

The Avro decoding client in `interchange` keeps around a `Row` packer, so that it can right-size allocations. This is just an optimization; semantically, we should use a new packer at the beginning of each call to `decode`. Normally things worked fine because the call to `finish_and_reuse` clears out the `Row`. However, in the error case, we return early without calling that function, and so there could be partial state left over, causing the _next_ returned row to be invalid.